requirements.txt
coreapi==2.3.3
Django==2.2.3
djangorestframework==3.9.4
Jinja2==2.10.1
Pygments==2.4.2
PyYAML==5.1.1
coreapi-cli
